MSVC5 does like -1UL...  replaced with (unsigned long) -1L

+++ libraries/libldap/cldap.c
@@ -109,7 +109,7 @@ cldap_open( char *host, int port )
 		}
 	    }
 
-	    if ( (address = inet_addr( host )) == -1UL ) {
+	    if ( (address = inet_addr( host )) == (unsigned long) -1L ) {
 		if ( (hp = gethostbyname( host )) == NULL ) {
 		    errno = EHOSTUNREACH;
 		    continue;
